app/requirements.txt
flask
flask_cors
python-dotenv
pymssql
redis
flask_jwt_extended
celery
kombu
ffmpeg-python
Gunicorn